mapbox to a react component
