+++ src/partials/_RestaurantDetails.scss
@@ -123,7 +123,7 @@
   }
 }
 
-.mapDetail {
+.mapContainer {
   height:400px;
 }
